refactor: split menu appearance resolver
