+++ Sources/CodexBar/StatusItemController+Menu.swift
@@ -990,17 +990,6 @@ extension StatusItemController {
         }
     }
 
-    static func systemMenuAppearanceName(
-        interfaceStyle: String?,
-        increaseContrast: Bool) -> NSAppearance.Name
-    {
-        let isDark = interfaceStyle?.caseInsensitiveCompare("Dark") == .orderedSame
-        if increaseContrast {
-            return isDark ? .accessibilityHighContrastDarkAqua : .accessibilityHighContrastAqua
-        }
-        return isDark ? .darkAqua : .aqua
-    }
-
     private func makeProviderSwitcherItem(
         providers: [UsageProvider],
         includesOverview: Bool,
